requirements.txt
pandas
numpy
streamlit
sentence-transformers
faiss-cpu
plotly
requests
arxiv